% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaCluster.R
\name{CopulaCluster}
\alias{CopulaCluster}
\title{Clustering by Matrix Features}
\usage{
CopulaCluster(FeatureMatrix, nClust)
}
\arguments{
\item{FeatureMatrix}{n*p features matrix with Patient names as rownames}

\item{nClust}{number of clusters}
}
\value{
clustering labels
}
\description{
This function cluster patients by extracted matrix features.
}
