% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaMatrix.R
\name{CopulaMatrix}
\alias{CopulaMatrix}
\title{Scatter data to Copula}
\usage{
CopulaMatrix(x1, x2, nGrid)
}
\arguments{
\item{x1}{Value array of first modality}

\item{x2}{Value array of second modality}
}
\value{
An discrete copula matrix with the size of nGrid*nGrid
}
\description{
This function tranforms orginal joint distribution to Copula and extract discret co-occurent matrix.
}
