% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix2Features.R
\name{Matrix2Features}
\alias{Matrix2Features}
\title{Matrix Statistics}
\usage{
Matrix2Features(dataMatrix)
}
\arguments{
\item{dataMatrix}{co-occurent matrix}
}
\value{
a dataframe of co-occurent matrix features
}
\description{
This function computes statistics features of an input Co-occurence Matrix.
Features are adapted from matlab toolbox RADIOMICS \url{https://github.com/mvallieres/radiomics/}.
}
\references{
[1] Haralick, R. M., Shanmugam, K., & Dinstein, I. (1973). Textural features for image classification. IEEE Transactions on Systems, Man and Cybernetics, smc 3(6), 610–621.

[2] Assefa, D., Keller, H., Ménard, C., Laperriere, N., Ferrari, R. J., & Yeung, I. (2010). Robust texture features for response monitoring of glioblastoma multiforme on T1 -weighted and T2 -FLAIR MR images: A preliminary investigation in terms of identification and segmentation. Medical Physics, 37(4), 1722–1736.

[3] Thibault, G. (2009). Indices de formes et de textures: de la 2D vers la 3D. Application au classement de noyaux de cellules. PhD Thesis, Université AIX-Marseille: p.172.

[4] Aerts, H.J.W.L. et al. Decoding tumour phenotype by noninvasive imaging using a quantitative radiomics approach. Nat. Commun. 5:4006 doi: 10.1038/ncomms5006 (2014).
}
